requirements.txt
numpy==1.14.3
pandas==0.24.2
Pillow==5.1.0
scikit-learn==0.21.3
scikit-image==0.13.1
scipy==1.1.0
torch==1.1.0
terminaltables==3.1.0
tensorflow-gpu==1.4.0
tensorflow-tensorboard=0.4.0
opencv-python==4.0.0.21
